TLB and Cache Search Worked out
stat tracking not yet fully integrated

+++ VirtualMemorySimulator/VirtualMemorySimulator/CacheFieldParser.cs
@@ -84,10 +84,10 @@ namespace VirtualMemorySimulator
             uint block_offs_tmp = block.block_offset;
             uint set_tmp = ((uint)block.set << 6);
             uint block_mask = (uint)((1 << (tagWidth)) - 1);
-            uint block_tag_tmp = (uint)((block.tag & block_mask) << (6 + SetIdxWidth));
+            ulong block_tag_tmp = (ulong)((ulong)(block.tag & block_mask) << (6 + SetIdxWidth));
 
-            uint retVal36 = block_tag_tmp | set_tmp | block_offs_tmp;
-            uint retVal24 = (retVal36 >> 12) & 0x0FFFFFF; 
+            ulong retVal36 = block_tag_tmp | set_tmp | block_offs_tmp;
+            uint retVal24 = (uint)(retVal36 >> 12) & 0x0FFFFFF; 
             ushort retVal12 = (ushort)(retVal36 & 0x0FFF);
 
             return new Tuple<uint,ushort>(retVal24, retVal12);
